Rectify latest cairo SVG export
It seems Cairo got updated and its SVG export is now broken in a new and
exciting way. Now it exports documents with dimensionless w/h. The issue
with that is that Cairo thinks that means point @ 72 pt / inch, but
everyone else including the actual svg spec interpret that as raw document
units or pixels, AFAICT completely breaking dpi scaling in common
viewers. Since we have to mess with Cairo's SVG files anyway, we just
fix this as well by re-writing the broken w/h into physical units
according to a hardcoded conversion factor that matches cairo's
hardcoded scale.