Make sure we asterisk-terminate all G0x commands.
While this is common in the wild, not terminating them violates the
spec. It also breaks JLCPCB pretty badly. It seems their human review
process uses a Gerber viewer that like most can handle this, and won't
notice anything out of the ordinary, but then their photoplotter chokes
on this and literally stops plotting the file, discarding anything that
is after that line. This error is then apparently ignored and the
resulting broken boards shipped to the customer.